hystart: fix ACK round split
The intuition decribed in "Taming the Elephants" (Ha et al.) would
suggest that ACK rounds/trains are split when now(snd_una) reaches
then(snd_nxt). The sum of ACK IATs in one round can then give us
an estimate for the time it took all packets in that round to
transfer over some bottleneck. We guess the bottleneck is filled to
the bandwidth-delay product when train_length >= RTT/2, because we
assume that RTT/2 = OWD.

However, current implementations of HyStart in tcp_cdg and tcp_cubic
split ACK rounds when now(snd_una) is greater than then(snd_nxt),
i.e., "start a new round when more than the currently sent data
is ACKed." This is counter-intuitive, because the last ACK received
for round n should really have been the first ACK of round n+1.

Reported-by: Veaceslav Roman <veaceslav.roman@orange.md>
Signed-off-by: Kenneth Klette Jonassen <kennetkl@ifi.uio.no>